init(base install): удалил файлы из мастер создал проект заново
