--- Api/Controllers/UserController.cs
@@ -1,30 +1,0 @@
-﻿using Api.Models;
-using Api.Services;
-using AutoMapper;
-using AutoMapper.QueryableExtensions;
-using DAL;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-
-namespace Api.Controllers
-{
-    [Route("api/[controller]/[action]")]
-    [ApiController]
-    public class UserController : ControllerBase
-    {
-        private readonly UserService _userService;
-
-        public UserController(UserService userService)
-        {
-            _userService = userService;
-        }
-
-        [HttpPost]
-        public async Task CreateUser(CreateUserModel model)=> await _userService.CreateUser(model);
-
-        [HttpGet]
-        public async Task<List<UserModel>> GetUsers() => await _userService.GetUsers();
-        
-    }
-}
