--- Api/Models/CreateUserModel.cs
@@ -1,28 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Api.Models
-{
-    public class CreateUserModel
-    {
-        [Required]
-        public string Name { get; set; }
-        [Required]
-        public string Email { get; set; } 
-        [Required]
-        public string Password { get; set; }
-        [Required]
-        [Compare(nameof(Password))]
-        public string RetryPassword { get; set; } 
-        [Required]
-        public DateTimeOffset BirthDate { get; set; }
-
-        public CreateUserModel(string name, string email, string password, string retryPassword, DateTimeOffset birthDate)
-        {
-            Name = name;
-            Email = email;
-            Password = password;
-            RetryPassword = retryPassword;
-            BirthDate = birthDate;
-        }   
-    }
-}
