--- Api/Program.cs
@@ -1,55 +1,0 @@
-using Api;
-using Api.Services;
-using Microsoft.EntityFrameworkCore;
-
-internal class Program
-{
-    private static void Main(string[] args)
-    {
-        var builder = WebApplication.CreateBuilder(args);
-
-        // Add services to the container.
-
-        builder.Services.AddControllers();
-        // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-        builder.Services.AddEndpointsApiExplorer();
-        builder.Services.AddSwaggerGen();
-
-        builder.Services.AddDbContext<DAL.DataContext>(options =>
-        {
-            options.UseNpgsql(builder.Configuration.GetConnectionString("PostgreSql"), sql => { });
-        });
-
-        builder.Services.AddAutoMapper(typeof(MapperProfile).Assembly);
-
-        builder.Services.AddScoped<UserService>();
-
-
-        var app = builder.Build();
-
-
-        using (var serviceScope = ((IApplicationBuilder)app).ApplicationServices.GetService<IServiceScopeFactory>()?.CreateScope())
-        {
-            if (serviceScope != null)
-            {
-                var context = serviceScope.ServiceProvider.GetRequiredService<DAL.DataContext>();
-                context.Database.Migrate();
-            }
-        }
-
-        // Configure the HTTP request pipeline.
-        //if (app.Environment.IsDevelopment())
-        {
-            app.UseSwagger();
-            app.UseSwaggerUI();
-        }
-
-        app.UseHttpsRedirection();
-
-        app.UseAuthorization();
-
-        app.MapControllers();
-
-        app.Run();
-    }
-}
