--- Api/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "PostgreSql": "Server=localhost;Port=5432;Database=Study2022db;User Id=dduser;Password=dduser"
-  }
-}
