--- DAL/Entities/User.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DAL.Entities
-{
-    public class User
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; } = "empty";
-        public string Email { get; set; } = "empty";
-        public string PasswordHash { get; set; } = "empty"; 
-        public DateTimeOffset BirthDate { get; set; }
-    }
-}
