--- DDStudy2022.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32929.385
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{9D768D4A-F102-44C8-A670-A2962066F7D9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DAL", "DAL\DAL.csproj", "{8C82853C-7844-4832-81BC-CDB4A4218FA6}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "Common\Common.csproj", "{EC4C89AE-F099-43C7-86F1-E3608A1CB1D1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9D768D4A-F102-44C8-A670-A2962066F7D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9D768D4A-F102-44C8-A670-A2962066F7D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9D768D4A-F102-44C8-A670-A2962066F7D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9D768D4A-F102-44C8-A670-A2962066F7D9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8C82853C-7844-4832-81BC-CDB4A4218FA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8C82853C-7844-4832-81BC-CDB4A4218FA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8C82853C-7844-4832-81BC-CDB4A4218FA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8C82853C-7844-4832-81BC-CDB4A4218FA6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EC4C89AE-F099-43C7-86F1-E3608A1CB1D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EC4C89AE-F099-43C7-86F1-E3608A1CB1D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EC4C89AE-F099-43C7-86F1-E3608A1CB1D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EC4C89AE-F099-43C7-86F1-E3608A1CB1D1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {918EB6F4-9693-43B5-833F-189FE8920A4C}
-	EndGlobalSection
-EndGlobal
